factorzation: biscuit extends unverifiedbiscuit
