+++ src/main/java/com/clevercloud/biscuit/token/Biscuit.java
@@ -2,35 +2,26 @@ package com.clevercloud.biscuit.token;
 
 import com.clevercloud.biscuit.crypto.KeyPair;
 import com.clevercloud.biscuit.crypto.PublicKey;
-import com.clevercloud.biscuit.datalog.*;
-import com.clevercloud.biscuit.error.FailedCheck;
-import com.clevercloud.biscuit.error.LogicError;
-import com.clevercloud.biscuit.token.format.SerializedBiscuit;
+import com.clevercloud.biscuit.datalog.SymbolTable;
 import com.clevercloud.biscuit.error.Error;
-
+import com.clevercloud.biscuit.token.format.SerializedBiscuit;
 import com.clevercloud.biscuit.token.format.SignedBlock;
 import io.vavr.control.Either;
 import io.vavr.control.Option;
-
-import static io.vavr.API.Right;
 
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.security.SignatureException;
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.ArrayList;
+import java.util.Base64;
+import java.util.Collections;
+import java.util.List;
 
 /**
  * Biscuit auth token
  */
-public class Biscuit {
-    final Block authority;
-    final List<Block> blocks;
-    final SymbolTable symbols;
-    final Option<SerializedBiscuit> container;
-    final List<byte[]> revocation_ids;
-
+public class Biscuit extends UnverifiedBiscuit {
     /**
      * Creates a token builder
      * <p>
@@ -100,11 +91,7 @@ public class Biscuit {
     }
 
     Biscuit(Block authority, List<Block> blocks, SymbolTable symbols, Option<SerializedBiscuit> container, List<byte[]> revocation_ids) {
-        this.authority = authority;
-        this.blocks = blocks;
-        this.symbols = symbols;
-        this.container = container;
-        this.revocation_ids = revocation_ids;
+        super(authority, blocks, symbols, container, revocation_ids);
     }
 
     /**
@@ -316,138 +303,6 @@ public class Biscuit {
                 this.container.get().proof.secretKey.isEmpty();
     }
 
-    Either<Error, World> generate_world() {
-        World world = new World();
-
-        for (Fact fact : this.authority.facts) {
-            world.add_fact(fact);
-        }
-
-        for (Rule rule : this.authority.rules) {
-            world.add_rule(rule);
-        }
-
-        for (int i = 0; i < this.blocks.size(); i++) {
-            Block b = this.blocks.get(i);
-
-            for (Fact fact : b.facts) {
-                world.add_fact(fact);
-            }
-
-            for (Rule rule : b.rules) {
-                world.add_rule(rule);
-            }
-        }
-
-        List<RevocationIdentifier> revocation_ids = this.revocation_identifiers();
-        long rev = symbols.get("revocation_id").get();
-        for (int i = 0; i < revocation_ids.size(); i++) {
-            byte[] id = revocation_ids.get(i).getBytes();
-            world.add_fact(new Fact(new Predicate(rev, Arrays.asList(new Term.Integer(i), new Term.Bytes(id)))));
-        }
-
-        return Right(world);
-    }
-
-    HashMap<String, Set<Fact>> check(SymbolTable symbols, List<Fact> ambient_facts, List<Rule> ambient_rules,
-                                     List<Check> authorizer_checks, HashMap<String, Rule> queries) throws Error {
-        Either<Error, World> wres = this.generate_world();
-
-        if (wres.isLeft()) {
-            Error e = wres.getLeft();
-            throw e;
-        }
-
-        World world = wres.get();
-
-        for (Fact fact : ambient_facts) {
-            world.add_fact(fact);
-        }
-
-        for (Rule rule : ambient_rules) {
-            world.add_rule(rule);
-        }
-
-        world.run(symbols);
-
-        ArrayList<FailedCheck> errors = new ArrayList<>();
-        for (int j = 0; j < this.authority.checks.size(); j++) {
-            boolean successful = false;
-            Check c = this.authority.checks.get(j);
-
-            for (int k = 0; k < c.queries().size(); k++) {
-                Set<Fact> res = world.query_rule(c.queries().get(k), symbols);
-                if (!res.isEmpty()) {
-                    successful = true;
-                    break;
-                }
-            }
-
-            if (!successful) {
-                errors.add(new FailedCheck.FailedBlock(0, j, symbols.print_check(this.authority.checks.get(j))));
-            }
-        }
-
-        for (int j = 0; j < authorizer_checks.size(); j++) {
-            boolean successful = false;
-            Check c = authorizer_checks.get(j);
-
-            for (int k = 0; k < c.queries().size(); k++) {
-                Set<Fact> res = world.query_rule(c.queries().get(k), symbols);
-                if (!res.isEmpty()) {
-                    successful = true;
-                    break;
-                }
-            }
-
-            if (!successful) {
-                errors.add(new FailedCheck.FailedAuthorizer(j + 1, symbols.print_check(authorizer_checks.get(j))));
-            }
-        }
-
-        for (int i = 0; i < this.blocks.size(); i++) {
-            Block b = this.blocks.get(i);
-
-            for (int j = 0; j < b.checks.size(); j++) {
-                boolean successful = false;
-                Check c = b.checks.get(j);
-
-                for (int k = 0; k < c.queries().size(); k++) {
-                    Set<Fact> res = world.query_rule(c.queries().get(k), symbols);
-                    if (!res.isEmpty()) {
-                        successful = true;
-                        break;
-                    }
-                }
-
-                if (!successful) {
-                    errors.add(new FailedCheck.FailedBlock(i + 1, j, symbols.print_check(b.checks.get(j))));
-                }
-            }
-        }
-
-        HashMap<String, Set<Fact>> query_results = new HashMap<>();
-        for (String name : queries.keySet()) {
-            Set<Fact> res = world.query_rule(queries.get(name), symbols);
-            query_results.put(name, res);
-        }
-
-        if (errors.isEmpty()) {
-            return query_results;
-        } else {
-            throw new Error.FailedLogic(new LogicError.Unauthorized(new LogicError.MatchedPolicy.Allow(0), errors));
-        }
-    }
-
-    /**
-     * Creates a Block builder
-     *
-     * @return
-     */
-    public com.clevercloud.biscuit.token.builder.Block create_block() {
-        return new com.clevercloud.biscuit.token.builder.Block(1 + this.blocks.size(), new SymbolTable(this.symbols));
-    }
-
     /**
      * Generates a new token from an existing one and a new block
      *
@@ -498,42 +353,6 @@ public class Biscuit {
         return new Biscuit(copiedBiscuit.authority, blocks, symbols, Option.some(container), revocation_ids);
     }
 
-    public List<List<com.clevercloud.biscuit.datalog.Check>> checks() {
-        ArrayList<List<com.clevercloud.biscuit.datalog.Check>> l = new ArrayList<>();
-        l.add(new ArrayList<>(this.authority.checks));
-
-        for (Block b : this.blocks) {
-            l.add(new ArrayList<>(b.checks));
-        }
-
-        return l;
-    }
-
-    public List<RevocationIdentifier> revocation_identifiers() {
-        return this.revocation_ids.stream()
-                .map(e -> RevocationIdentifier.from_bytes(e))
-                .collect(Collectors.toList());
-    }
-
-    public List<Option<String>> context() {
-        ArrayList<Option<String>> res = new ArrayList<>();
-        if (this.authority.context.isEmpty()) {
-            res.add(Option.none());
-        } else {
-            res.add(Option.some(this.authority.context));
-        }
-
-        for (Block b : this.blocks) {
-            if (b.context.isEmpty()) {
-                res.add(Option.none());
-            } else {
-                res.add(Option.some(b.context));
-            }
-        }
-
-        return res;
-    }
-
     /**
      * Prints a token's content
      */
@@ -552,27 +371,6 @@ public class Biscuit {
         s.append("\t]\n}");
 
         return s.toString();
-    }
-
-    /**
-     * Default symbols list
-     */
-    static public SymbolTable default_symbol_table() {
-        SymbolTable syms = new SymbolTable();
-        syms.insert("authority");
-        syms.insert("ambient");
-        syms.insert("resource");
-        syms.insert("operation");
-        syms.insert("right");
-        syms.insert("current_time");
-        syms.insert("revocation_id");
-
-        return syms;
-    }
-
-    @Override
-    protected Object clone() throws CloneNotSupportedException {
-        return super.clone();
     }
 
     public Biscuit copy() throws Error {
